requirements.txt
Django==5.2.10
Pillow==12.1.0
django-allauth==65.4.1
